+++ packages/jest-jasmine2/package.json
@@ -18,8 +18,7 @@
     "jest-message-util": "^22.4.0",
     "jest-snapshot": "^22.4.0",
     "jest-util": "^22.4.1",
-    "pretty-format": "^22.4.0",
-    "source-map-support": "^0.5.6"
+    "pretty-format": "^22.4.0"
   },
   "devDependencies": {
     "jest-runtime": "^22.4.2"
