+++ packages/jest-jasmine2/src/index.js
@@ -18,7 +18,6 @@ import path from 'path';
 import fs from 'graceful-fs';
 import installEach from './each';
 import {getCallsite} from 'jest-util';
-import sourcemapSupport from 'source-map-support';
 import JasmineReporter from './reporter';
 import {install as jasmineAsyncInstall} from './jasmine_async';
 
@@ -120,36 +119,6 @@ async function jasmine2(
     runtime.requireModule(config.setupTestFrameworkScriptFile);
   }
 
-  const sourcemapOptions = {
-    environment: 'node',
-    handleUncaughtExceptions: false,
-    retrieveSourceMap: source => {
-      const sourceMaps = runtime.getSourceMaps();
-      const sourceMapSource = sourceMaps && sourceMaps[source];
-
-      if (sourceMapSource) {
-        try {
-          return {
-            map: JSON.parse(fs.readFileSync(sourceMapSource)),
-            url: source,
-          };
-        } catch (e) {}
-      }
-      return null;
-    },
-  };
-
-  // For tests
-  runtime
-    .requireInternalModule(
-      require.resolve('source-map-support'),
-      'source-map-support',
-    )
-    .install(sourcemapOptions);
-
-  // For runtime errors
-  sourcemapSupport.install(sourcemapOptions);
-
   if (globalConfig.enabledTestsMap) {
     env.specFilter = spec => {
       const suiteMap =
@@ -162,16 +131,12 @@ async function jasmine2(
     env.specFilter = spec => testNameRegex.test(spec.getFullName());
   }
 
-  try {
-    runtime.requireModule(testPath);
-    await env.execute();
+  runtime.requireModule(testPath);
+  await env.execute();
 
-    const results = await reporter.getResults();
+  const results = await reporter.getResults();
 
-    return addSnapshotData(results, snapshotState);
-  } finally {
-    sourcemapSupport.resetRetrieveHandlers();
-  }
+  return addSnapshotData(results, snapshotState);
 }
 
 const addSnapshotData = (results, snapshotState) => {
