+++ packages/jest-runner/package.json
@@ -18,6 +18,7 @@
     "jest-runtime": "^22.4.2",
     "jest-util": "^22.4.1",
     "jest-worker": "^22.2.2",
+    "source-map-support": "^0.5.6",
     "throat": "^4.0.0"
   }
 }
